Endpoint created based by score

+++ dev/healthrecords/src/main/java/devzone/healthrecords/repository/ClientRepository.java
@@ -29,6 +29,6 @@ public interface ClientRepository extends JpaRepository<Client, Long> {
 	+"FROM client c, health_issue h "
 	+ "WHERE c.id = h.client_id AND c.id= :id"
 	, nativeQuery = true)
-	int sumDegree(@Param("id") Long id);
-	
+	double sumDegree(@Param("id") Long id);
+
 }
